bff: place all integrations translators in one file
